divided tests by class
